анализ стран World Bank/requirements.txt
matplotlib == 3.8.4
numpy == 1.26.4
pandas == 2.2.2
plotly == 5.22.0
requests == 2.32.2
seaborn == 0.13.2
SQLAlchemy == 2.0.30
scikit-learn == 1.4.2
scipy == 1.13.1
python-dotenv == 0.21.0
